Update develop branch with latest commits from 2.1.0

+++ src/test/script_P2SH_tests.cpp
@@ -317,7 +317,7 @@ BOOST_AUTO_TEST_CASE(AreInputsStandard)
     txFrom.vout[6].scriptPubKey = GetScriptForDestination(CScriptID(twentySigops));
     txFrom.vout[6].nValue = 6000;
 
-    AddCoins(coins, txFrom, 0);
+    AddCoins(coins, txFrom, 0, uint256());
 
     CMutableTransaction txTo;
     txTo.vout.resize(1);
